created project

+++ UserCRUDApi/Infractructure/Interfaces/IRepositoryBase.cs
@@ -6,7 +6,7 @@ namespace UserCRUDApi.Infractructure.Interfaces;
 public interface IRepositoryBase<T> where T : BaseModel
 {
     DbSet<T> GetSet();
-    ValueTask<T?> GetByIdAsync(int id);
+    ValueTask<T?> GetByIdAsync(long id);
     ValueTask<T> CreateAsync(T entity);
     ValueTask<T> DeleteAsync(T entity);
     ValueTask<T> UpdateAsync(T entity);
